refactor(structure): skip lazy-loading of default pane types
