+++ dev/test-studio/structure/resolveStructure.ts
@@ -231,6 +231,7 @@ export const structure: StructureResolver = (S, {schema, documentStore, i18n}) =
                     title: 'Authors & Books',
                     options: {
                       filter: '_type == "author" || _type == "book"',
+                      apiVersion: '2023-07-28',
                     },
                   }),
                 ),
@@ -377,6 +378,43 @@ export const structure: StructureResolver = (S, {schema, documentStore, i18n}) =
               S.divider(),
 
               S.documentTypeListItem('sanity.imageAsset').title('Images').icon(ImagesIcon),
+            ]),
+        ),
+
+      S.listItem()
+        .title('Structure pane types')
+        .id('structure-pane-types')
+        .child(
+          S.list()
+            .id('structure-pane-types')
+            .title('Structure pane types')
+            .items([
+              S.listItem()
+                .title('List')
+                .child(
+                  S.list()
+                    .title('List')
+                    .items([
+                      S.listItem().title('Static item 1').id('static-item-1'),
+                      S.listItem().title('Static item 2').id('static-item-2'),
+                    ]),
+                ),
+              S.listItem()
+                .title('Document list')
+                .child(
+                  S.documentList()
+                    .title('Document list')
+                    .filter('_type == "author"')
+                    .apiVersion('2023-07-28'),
+                ),
+              S.listItem()
+                .id('doc-item-singleton')
+                .title('Document')
+                .child(S.document().documentId('grrm').schemaType('author')),
+              S.listItem()
+                .id('grrm')
+                .title('Component')
+                .child(S.component(JsonDocumentDump).id('json-dump').title('Custom component')),
             ]),
         ),
