+++ packages/sanity/src/structure/panes/document/DocumentPane.tsx
@@ -17,7 +17,7 @@ import {usePaneRouter} from '../../components'
 import {structureLocaleNamespace} from '../../i18n'
 import {type DocumentPaneNode} from '../../types'
 import {ErrorPane} from '../error'
-import {LoadingPane} from '../loading'
+import {LoadingPane} from '../loading/LoadingPane'
 import {CommentsWrapper} from './comments'
 import {useDocumentLayoutComponent} from './document-layout'
 import {DocumentPaneProvider} from './DocumentPaneProvider'
