--- packages/sanity/src/structure/panes/list/index.ts
@@ -1,1 +1,0 @@
-export {ListPane as default} from './ListPane'
