--- packages/sanity/src/structure/panes/loading/index.ts
@@ -1,1 +1,0 @@
-export {LoadingPane} from './LoadingPane'
