--- packages/sanity/src/structure/panes/unknown/index.ts
@@ -1,1 +1,0 @@
-export * from './UnknownPaneType'
