--- packages/sanity/src/structure/panes/userComponent/index.ts
@@ -1,1 +1,0 @@
-export {UserComponentPane as default} from './UserComponentPane'
